correcting some warnings with visual studio 2015

+++ modules/ustk_core/src/RFConversion/usRFToPreScan2DConverter.cpp
@@ -180,7 +180,7 @@ void usRFToPreScan2DConverter::convert(const usImageRF2D<short int> &rfImage, us
   int k = 0;
   for (int i = 0; i < h; i+=m_decimationFactor) {
     for (int j = 0; j < w ; ++j) {
-      unsigned int  vcol = ((comp[i + h * j] - min) / maxMinDiff) * 255;
+      unsigned int  vcol = (unsigned int) (((comp[i + h * j] - min) / maxMinDiff) * 255);
       preScanImage[k][j] = (vcol>255)?255:vcol;
     }
     k++;
